refactor: 파일 구조 변경 -> ResponseDto 하위 패키지 생성
