+++ Server/src/main/java/JGS/CasperEvent/domain/event/dto/ResponseDto/lotteryEventResponseDto/CasperBotResponseDto.java
@@ -1,4 +1,4 @@
-package JGS.CasperEvent.domain.event.dto.ResponseDto;
+package JGS.CasperEvent.domain.event.dto.ResponseDto.lotteryEventResponseDto;
 
 import JGS.CasperEvent.domain.event.entity.casperBot.CasperBot;
 import org.springframework.data.annotation.Id;
