+++ Server/src/main/java/JGS/CasperEvent/domain/event/dto/ResponseDto/rushEventResponseDto/MainRushEventResponseDto.java
@@ -1,4 +1,4 @@
-package JGS.CasperEvent.domain.event.dto.ResponseDto;
+package JGS.CasperEvent.domain.event.dto.ResponseDto.rushEventResponseDto;
 
 import JGS.CasperEvent.domain.event.entity.event.RushEvent;
 import lombok.Getter;
