+++ Server/src/main/java/JGS/CasperEvent/domain/event/dto/ResponseDto/rushEventResponseDto/ResultRushEventOptionResponseDto.java
@@ -1,4 +1,4 @@
-package JGS.CasperEvent.domain.event.dto.ResponseDto;
+package JGS.CasperEvent.domain.event.dto.ResponseDto.rushEventResponseDto;
 
 public record ResultRushEventOptionResponseDto(String mainText, String resultMainText, String resultSubText) {
     public static ResultRushEventOptionResponseDto of(RushEventOptionResponseDto rushEventOptionResponseDto) {
