+++ Server/src/main/java/JGS/CasperEvent/domain/event/service/redisService/RedisService.java
@@ -1,6 +1,6 @@
 package JGS.CasperEvent.domain.event.service.redisService;
 
-import JGS.CasperEvent.domain.event.dto.ResponseDto.CasperBotResponseDto;
+import JGS.CasperEvent.domain.event.dto.ResponseDto.lotteryEventResponseDto.CasperBotResponseDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
